Reduce memory overhead of ant runtests.
The <junitreport> dumps out XML files containing test results, and
containing the STDOUT&STDERR for every tests.  These are parsed into
in-memory DOMs that are passed to an XSLT processor to generate the
HTML reports.

The amount of logging trace our tests generate dominates the amount of
memory taken by this post-processing step.

This change cuts out roughly 9/10ths of that.
Verbose logging can be reenabled via the junit.verbose System property
  ant -Djunit.verbose=true runtests