Split out the bits of JavaScript rendering code relating to deciding whether spaces or newlines are required/allowed between tokens.
Ihab pointed out that the side-by-side renderer is splitting lines in
ways that allow for semicolons to be inserted after return and throw.

This consolidates the code related to that.  In another CL I will
rewrite existing renderers to use this consistently, and in a third, I
will rewrite JsPrettyPrinter to do better formatting.  This CL is
simple and semantics preserving.  The second will be a more
substantial change but semantics preserving.  The third will not be
semantics preserving.