Modified parser to normalize identifiers using Unicode normal form C (NFC).
ES3 and ES3.1 assume that the input is in NFC, but do not require it.
This change normalizes identifiers and unquoted object constructor
property names, but does not modify quoted constructor property names.
The changes to browser-expectations.html verify that this
normalization will not cause confusion in existing browsers.

Both versions of ES also have confusion around comments in restricted
productions.  As discussed in "Multi-line comments and semicolon
insertion" on es3.x-discuss@mozilla.org and
  (function () { a: for (;;) { for (;;) { break/*
  */a; } return 1; } return 2; })()
is implemented inconsistently.  This change makes sure that our parser
correctly handles the interaction between mult-line comments and
semicolon insertion.
Another pending change makes sure that our renderers will not emit a
multiline comment where it is not allowed.