Issue 8698  http://codereview.appspot.com/8698 Fix IE bugs in DOMita
TamedXMLHttpRequest was failing because XMLHttpRequest does not exist
in IE 6 and is instead implemented as an ActiveXObject.

tameNode was failing in IE because colons cannot be part of a custom
property attached to a <TABLE> element.

tameNode was also failing on text and comment nodes since those can't
have custom properties at all.

R=ihab.awad@gmail.com,erights@gmail.com