Issue 8704  http://codereview.appspot.com/8704 Checks that E4X constructs are blocked by our parser.
E4X (EcmaScript for XML) is a JS variant that allows for inline XML
blobls.  This checks that examples of E4X will not pass our parser,
and that when we see a script tag with an E4X mime-type, that we do
not emit a script tags with an E4X mime-type.

R=metaweta@gmail.com