Issue 12509  http://codereview.appspot.com/12509 Issue 955:  	 Parity error: Unowned property
In testbed.js:

        if (!name && ('name' in ctor) && !___.hasOwnProp(ctor, 'name')) {
          name = ctor.name;
        }

appears in a context where ctor is a genuine constructor function. This
will only be true when Function.prototype has a 'name' property (as it does
at least on FF, where it is the empty string) and ctor does not override it
with its own 'name' property. I suspect you meant not to "!" the hasOwnProp
test.

R=erights@gmail.com