Issue 13042  http://codereview.appspot.com/13042 Fix svn:mime-type inference in myvn
Changes the bits of tools/myvn filetypes that decide when to set the
mime types, and the file extension to mime-type mapping.

svn add examines each file, and assumes that any file containing a
byte with the high-bit set is a binary file.  It records this decision
by setting the property svn:mime-type to application/octet-stream.

Previously myvn filetypes would only set the mime-type for files
without any which missed all images.

R=jasvir@gmail.com