Issue 21066  http://codereview.appspot.com/21066 Issue 986: Domita needs an implementation of element.contains()
From http://code.google.com/p/google-caja/issues/detail?id=986:
    YUI uses el.contains(node) to check if one DOMNode is contained
    within another.  The contains method was introduced in IE.
    compareDocumentPosition is specified in DOM Level 3 and *it* can
    be used to emulate contains() in bridal as described here
    http://www.quirksmode.org/blog/archives/2006/01/contains_for_mo.html
    but domita doesn't expose compareDocumentPosition.

If document.body.contains exists, add that to TameBackedNode.
If document.body.compareDocumentPosition exists, add that.
If the second exists but not the first, implement contains using
compareDocumentPosition as in quirksMode, but fixing it for the case
of el.contains(el).

R=jasvir@gmail.com