Issue 21046  http://codereview.appspot.com/21046 Issue 947: HTML emitter vs. getAttribute() inconsistent
Fix some inconsistencies in HtmlSanitizer and HtmlValidator.
This makes HtmlValidator more paranoid, and fixes the HTML attribute
schema to whitelist the accesskey attribute.

I tried repeating the issue in a number of revisions from the last 6
months, but couldn't.  But the schema fix is necessary, and the sanity
check can't hurt.

R=jasvir@gmail.com