Issue 32092  http://codereview.appspot.com/32092 Allow dynamic script loading via proxy
Modifies domita to allow creation and manipulation of script elements.

Cajoled code can now create a script element, set its "src" attribute to
a URL passed through the url policy with the mime-type text/javascript, and
add that script to the DOM.  It cannot modify the textual content of a
script element.

Outstanding questions: do we need to prevent replay attacks and if so how?
If code can get a handle to a script element from the container, it
can clone it and add it to the DOM, possibly replaying script already
executed by the container.

Still TODO: http://code.google.com/p/google-caja/wiki/UrlPolicy

R=metaweta@gmail.com