Issue 53101  http://codereview.appspot.com/53101 Funnel output from renderers into a Concatenator.
Ihab and Mark realized that it's easier to do things like renderers
with original source line mappings if all the renderers bottom out in
a Concatenator instead of in an Appendable.

This changes the API to make that a fact, though the pretty and
minimal JS printers do not yet forward FilePositions through to the
Concatenator too.

R=erights@gmail.com