teach appspot.py about git branches http://codereview.appspot.com/90069
Right now, appspot.py keeps state in a file named .appspot-change,
which is fine when you have a separate svn checkout for each change,
but doesn't work so well if you're putting changes in git topic branches
in a combined git/svn workdir.

This changes appspot.py so that when it notices you're using git,
it stores state in .appspot-change.$GITBRANCH instead.

R=ihab.awad@gmail.com