cajole_html is refusing <a href="#foo"> http://codereview.appspot.com/88149 http://code.google.com/p/google-caja/issues/detail?id=1025
cajole_html uses a unique PluginEnvironment
that allows limited filesystem references,
but barfs on most uses of #fragments.

this change relaxes the restriction on #fragments.
so now <a href="#foo"> is allowed,
and if href="foo.html" is allowed,
then href="foo.html#bar" is also allowed.

there's no testcase in this change because there
isn't already a test structure for any of this.
(emma says 0% coverage)
I might add tests later.

R=ihab.awad@gmail.com