Simplify Rule and Rewriters signatures http://codereview.appspot.com/101043
We were passing MessageQueues around in may places, throughout the various Rewriter implementations.  But we always pass the same MessageQueue around -- the one that is passed into the initial expand call.

I removed the MessageQueue parameter from various Rule and Rewriter methods and added a MessageQueue parameter to Rewriter.  Rule already points to a Rewriter so Rule's have access indirectly to the MessageQueue.

R=metaweta@gmail.com