reading from /Users/msamuel/work/google-caja-utd/.appspot-change
Clean up warnings for homepage test gadgets
http://codereview.appspot.com/116079

We generate a number of cajoled files that are included in iframes on
the caja home page to show how we stop certain attacks.
These have known bad properties, so I added an ignore attribute to the
output task so that "ant demos" does not display a bunch of spurious
warnings.

R=jasvir@gmail.com