An Alpha Renamer that makes it safe to embed user code in generated code. http://codereview.appspot.com/110096
Returns an expression that is structurally the same as e, but with
identifiers renamed using the alpha renaming.

If the input expression contains synthetic references or declarations, then
those synthetic identifiers will not be rewritten.

So after alpha renaming, matching references and declarations with
non-synthetic identifiers is trivial.

This renaming does not affect property names, so any relationship between
a property name and a global variable will be broken.  For this reason,
all free variables must be specified in the input context or
freeSynthetics set.

R=ihab.awad@gmail.com