Fixed NPE on a:link splitting. http://codereview.appspot.com/122048
I was running the example
  bin/cajole_html <(echo '<div style="color:\"></div>')

This failed with a null pointer exception because
CssRewriter.rewriteHistorySensitiveRuleset was lazily creating lists
in a widgy way.

I also changed PluginCompilerMain and Config so that I can run
cajole_html from the command line as in the example above.
Previously, it objected to named pipes.

R=jasvir@gmail.com