Fixes issue 1052: fastpath canSet test is inconsistent http://codereview.appspot.com/124103
The fastpath test for property assignment was just testing whether the
fastpath flag is truthy, which it might be for an inherited writable
property. Instead, we now test that its value is the object itself (as
the slow path test was already doing), in order to test that the own
property is writable.

R=benl@google.com