Some cleanup http://codereview.appspot.com/135052
Made ordering of script loading in domita_test consistent with that in
domita-minified.js, and made sure domita-minified includes the html
sanitizer, which is now required by the html-emitter because of
document.write support.

Fixed minor problem with SafeIdentifierMaker -- the name "eval" has
special significance, so make sure it doesn't output it.  This is
unlikely to have happened in practice, since the code is unused, and
would only surface if run over a file with more than 5 * pow(26, 3)
identifiers.

Added a consistency check to FunctionDeclaration.

Cleaned up uses of feature detection in bridal.js.

Added an iterators utility class to make it easy to implement
SafeIdentifierMaker as a filter operation over a stream of candidate
names.

R=metaweta@gmail.com