Fix rendering of Conditionals http://codereview.appspot.com/154078
It is possible to programatically construct a parse tree containing
conditionals where the rendered form attaches an else to the wrong
conditional.

For e.g., it is possible to programatically move
  if (foo) bar();
into the body of
  if (baz) @body else boo();
which should result in the structure
  if (baz) { if (foo) bar(); } else boo();
but which naively renders to
  if (baz) if (foo) bar(); else boo();
which is equivalent to
  if (baz) { if (foo) bar(); else boo(); }

Our code does not do this right now -- we are liberal with blocks --
so this does not constitute an attack vector.

This adds a method to statement, hasHangingConditional, and changes
the Conditional.render to add curly braces when the structure would
require them to resolve ambiguity.

Please look especially at the definition and use of
hasHangingConditional in Conditional.java and the tests in ParserTest.java

R=metaweta@gmail.com