Issue 1174 - <button>.cloneNode fails on IE http://codereview.appspot.com/165098
http://code.google.com/p/google-caja/issues/detail?id=1174

ie[67] doesn't let you change a button's type after it's created.
it's the same as the <input> type problem, you have to specify the
type in createElement.  bridal's cloneNode handles <input> correctly
but not <button.


This change fixes constructClone in bridal.js.  I also checked
html4-attributes-defs.json to see whether there are any other nodes
with a "type" or "value" attribute that need to be fixed up.  The only
other ones are PARAM and LI, UK, OL, and disallowed elements like
OBJECT and LINK, none of which are inputs.

R=felix8a@gmail.com