bad html attributes downgraded to a warning http://codereview.appspot.com/164070
Prior to r3844, html with an invalid attr like
  <div title='bob's foo'>
would have the bad "foo'" attribute deleted with a warning.

r3844 changed that to be a fatal error.

Bad attributes are unfortunately common, for many reasons.
Usually because they're typos of some sort.

This CL downgrades the error to a warning.

R=mikesamuel@gmail.com