Move test from DefaultValijaRewriterTest that are actually enforced in Cajita http://codereview.appspot.com/206081
One of the test helpers in DefaultValijaRewriterTest changes the
rewriter used.  This means that some tests were actually being treated
as cajita code.  A few of those are checkFails tests that only run the
valija transform which does not enforce the property.  I moved those
tests to Cajita and made sure that the widgy test helpers clean up
after themselves.

R=ihab.awad@gmail.com