Fix error trapping in embedded modules. http://codereview.appspot.com/216044
Cajita keeps track, in the module handler of the outcome of the last module.

The same module handler is used to handle the modules a module loads.
The lastOutcome variable is shared state across modules which means
that the outer modules result is not independent of any modules it
loads.

This fixes that by using a local variable inside the handle function
to store intermediate results.  The problem arises because lastOutcome
is set on both sides of the call to newModule.instantiate.

R=erights@gmail.com