Fix debug symbol render so it can be used with embedded rendering mode http://codereview.appspot.com/638043
Prior, the DefaultGadgetRewriter, which is used by the playground,
always rendered in non-debug mode.

Adds a "Debug" checkbox to the playground UI and threads that bit
through to the DefaultGadgetRewriter.

This flushed out a problem whereby
CajoledModule.renderWithDebugSymbols always rendered with
embeddableMode off even if the RenderContext had that flag set.
These changes should be semantics preserving where debug is false.

Changed CajoledModule.renderWithDebugSymbols to thread through the
normal RenderContext flags.  There were many methods that took an
output buffer and a Callback<IOException>.  I took the opportunity to
refactor those to take a RenderContext containing a Concatenator
instead.  Those signature changes should be semantics preserving.

R=ihab.awad@gmail.com