allow min-height:auto, etc. http://codereview.appspot.com/1030043
min-height:auto is a weirdness specific to IE7.
it's not in the CSS spec, it's not in the IE docs,
and all browsers other than IE7 will ignore the bad value.

Caja emits a warning and removes the css rule.

YUI uses min-height:auto as a targeted IE7 hack,
and I see a few instances of it elsewhere on the web.

this patch adds "auto" to the css whitelist,
so Caja will pass it through without complaint.

R=mikesamuel@gmail.com