Refactor object constructors http://codereview.appspot.com/1033042
This changes the structure of ObjectConstructors to make it easier to
support setters/getters, e.g. ({ foo: bar, get baz() { } }).

This still needs some work to fix ancillary tests, and to possibly
refactor object constructor rewriting to have a rule for
object properties instead of having a bunch of loops.

R=metaweta@gmail.com