Fix synthetic declaration grouping. http://codereview.appspot.com/1562041
This change is a bugfix separated out from a much larger CL : 207079
to address a problem that Michael John Ficarra has been running into.

FAILED TO REWRITE <<<var x0___, x1___, x2___>>>
java.lang.ClassCastException: com.google.caja.parser.js.Declaration cannot be
cast to com.google.caja.parser.js.ExpressionStmt
       at com.google.caja.parser.quasiliteral.CajitaRewriter$61.fire(CajitaRewriter.java:1975)
       at com.google.caja.parser.quasiliteral.Rewriter.expand(Rewriter.java:90)
       at com.google.caja.parser.quasiliteral.Rule.expandAllTo(Rule.java:169)
       at com.google.caja.parser.quasiliteral.Rule.expandAll(Rule.java:160)

The immediate problem is that CajitaRewriter uses two sets of rules.
One of those sets was changed to decompose declarations into an
initializer declaration added as a start statement to the scope and an
initializer.  The other, in SyntheticRuleSet, was not.

Although not manifesting in Michael John's report, there was another bug :
CajitaRewriter's multi declaration rule was assuming that the rewriting step
for an individual declaration within a multi-declaration always succeeded.

R=metaweta@gmail.com