Give cajita-module.js distinction between absolute module references and cajoled code references. http://codereview.appspot.com/1727041
cajita-module.js as it stands has a notion of a 'module identifier'
which is relative to the requester (in practice, a relative URL) and
'resolving' it to an absolute reference -- to the cajoled code.
Unfortunately, this means that if dynamic cajoling is used, then the
relative reference resolver has to destructure a cajoler-service URL,
which is Just Plain Wrong. It also tries to add a '.js' extension if
absent, which is awkward for various use cases.

This changes it to have three stages:
  1. A module identifier, which can be anything but is usually a
     relative URL, 
     
     is resolved (by the optionally host-page-supplied module identifier
     resolver -- this part hasn't changed) to 

  2. an absolute URL for the module (which usually refers to a .js 
     file),

     which is converted by the optional 'cajoler finder' function
     (defaulting to the caja.appspot.com service for easy playing, but
     could easily be adding .out.js instead) to

  3. a URL for cajoled code which may be <script> loaded.

Additionally:

  * The code was somewhat confusing or confused about whether a 'module
    ID' is a relative URL/not yet put through the resolver function or
    an absolute URL/has been through the resolver. A 'module ID' is now
    always pre-resolver: it is strictly the string passed to load().
    Absolute references are 'module URLs', and the moduleID property
    of module objects has been changed to reflect this.

  * The module identifier resolver now uses uri.js to perform RFC-3986-
    compliant relative URI resolution, rather than reimplementing it
    sloppily.

All the code which mentions cajita-module.js has been adjusted to match.
(Note that some of the demos don't work either after or before this
change.)

R=ihab.awad@gmail.com