Issue 1237: Playground runs script before cajoled HTML laid out. http://codereview.appspot.com/1666046
The offsetWidth,offsetHeight, etc. are only valid for HTML elements that are part of the DOM and laid out.  If any ancestor is display:none then it will not be laid out so offsetWidth,etc. return stale data.

This changes PlaygroundView to make the editor tab visible before the dynamic JS portion of the cajoled output is run.

R=jasvir@gmail.com