Shift to loading the Cajita runtime into an iframe rather than the host page. http://codereview.appspot.com/1812042
Added file caja.js which creates an iframe, then loads everything from Cajita to HostTools into that frame, and calls a callback with all that stuff so the host page can load gadgets and grant access to the host DOM. Changed the container/ demos to use it. Does not update the playground or any other demos.

R=ihab.awad@gmail.com