Fix DomitaTest valija testReadOnlyNotEditable to actually test. http://codereview.appspot.com/1896043
The documentRO variable was not actually visible in DomitaTest in Valija mode, because it was not marked readable on the outers object; therefore testReadOnlyNotEditable was getting its expected failures from documentRO being unbound rather than edits failing.

This change makes the test check that access through documentRO work, and makes domita_test.html create documentRO using cajita.setPub so that it is visible.

R=jasvir@gmail.com