Fixed pipeline generation in when only JS emitted by PluginCompilerMain http://codereview.appspot.com/2007046
There was logic in PluginCompilerMain that duplicated some of the
logic in Config.  I got rid of that and got rid of the -s flag that
Mike Stay pointed out was now redundant.  The onlyJsEmitted check
(previously based on -s) is now based on the presence of -j (JS output
file) and absence of -h (HTML output file).

This was causing problems because the duplicative and incorrect logic
in PluginCompilerMain was causing JS to appear as an output target
along with CAJOLED_MODULE_ONE, which is nonsensical, and caused the
ValidateJavaScriptStage to be scheduled before the HTML_TO_JS stage.

R=metaweta@gmail.com