Fixes Issue 1292: a NPE in the array index optimization stage http://codereview.appspot.com/2790042
Mark found a NPE in the array index optimization code.  It was
triggered by loading the game of life into the playground, but was not
actually related to the actual code loaded.  The triggering code was
some JS generated by the HTML compilation stage that used `this`.

This is not declared, so the ArrayIndexOptimizationStage scoping code
that assumed a name was declared before being read failed.

This fixes that issue and adds a minimal testcase.

R=erights@google.com