Replace package.html files with annotatable package-info.java. http://codereview.appspot.com/2817042
From 7.4 Package Declarations of the JLS 3ed:
http://java.sun.com/docs/books/jls/third_edition/html/packages.html

The manner in which this restriction is enforced must, of necessity,
vary from implementation to implementation. The following scheme is
strongly recommended for file-system-based implementations: The sole
annotated package declaration, if it exists, is placed in a source
file called package-info.java in the directory containing the source
files for the package. This file does not contain the source for a
class called package-info.java; indeed it would be illegal for it to
do so, as package-info is not a legal identifier. Typically
package-info.java contains only a package declaration, preceded
immediately by the annotations on the package. While the file could
technically contain the source code for one or more package-private
classes, it would be very bad form.  It is recommended that
package-info.java, if it is present, take the place of package.html
for javadoc and other similar documentation generation systems. If
this file is present, the documentation generation tool should look
for the package documentation comment immediately preceding the
(possibly annotated) package declaration in package-info.java. In this
way, package-info.java becomes the sole repository for package level
annotations and documentation. If, in future, it becomes desirable to
add any other package-level information, this file should prove a
convenient home for this information

R=jasvir@gmail.com