First pass at integrating the pipeline cache. http://codereview.appspot.com/2096041
Integrates job caching into the Cajoling pipeline so that we can avoid
recajoling jquery and other widely used libraries every time it
appears.

- Rewire script extracting in RewriteHtmlStage to use the
  placeholderIds in JobEnvelope.  This required follow-on changes to
  TemplateCompiler and SafeHtmlMaker.

- Added a PipelineFetchStage and PipelineStoreStage and wired them
  into the PipelineMaker.  The former replaces jobs that are in the cache
  with the processed result.  The latter caches the results of processing.

- Reworked CompileHtmlStage to pass through cached JS and CSS jobs,
  and to split output along cache key boundaries.

- Added a test suite of pipeline cache tests.


This change would be simpler if, instead of pulling JS out of HTML,
leaving placeholders in the HTML, we split the HTML around the
scripts.
I.e. instead of decomposeing
  HTML: <p>Hello, <script>alert(42)</script>, World</p>
into
  HTML: <p>Hello, <span phid="$1"></span>, World</p>
  JS:   alert(42)  // For placeholder $1
we represented HTML jobs as separately cachable ranges of nodes like
the selection model used in browsers to copy/paste HTML.
  HTML: |<p>Hello, |
  JS:   alert(42)
  HTML: |World</p>|

R=jasvir@gmail.com