workaround for Flash sizing problem http://codereview.appspot.com/4298057
the playground's Flash demo looks bad because the "natural" size of
the youtube video is larger than the dimensions given to embedSWF.
normally the video will resize itself to the bounding box, but
something about flashbridge causes that to not work.

I'm going to open a bug on that, but in the meantime this change
just makes the embed the expected size.

R=jasvir@gmail.com