fix playground clock demo http://codereview.appspot.com/4306048
the clock demo works fine, but if you run something after running
the clock demo, the js console starts spitting out errors every
second, because the clock demo set an interval handler that's still
firing every second.

afaict, there's no security implication to this leak, since the
interval handler is as sandboxed as it would be if it were just
one gadget on a page full of gadgets, but the unintentional
resource leak is annoying, and maybe caja should automatically
cancel timers when a gadget is unloaded.

in the meantime, this change makes the clock demo cancel the
interval handler when it notices it's been unloaded.

R=jasvir@gmail.com