run all sunspider benchmarks in playground http://codereview.appspot.com/4271069
this is a quick hack to run all the sunspider benchmarks in the
playground at once, basically just concatenating them all together
into a single file, with some renaming.

perhaps a better way to do this would be to modify the playground so
that it has a way to run N examples in sequence and accumulate the
result, but this was much easier to do.

the combined file is a little over 512K.
com.google.caja.lexer.FetchedData has a hard limit of 256K,
so I raised it to 1M.

R=jasvir@gmail.com