Fix escaping text span rendering in CDATA elements. http://codereview.appspot.com/4348049
We had a bug in Nodes.render around CDATA elements.
This was not exploitable because we are careful in our CSS and JS
renderers to not allow the sequences "<!--" or "-->" to appear in the
output, but this adds a sanity check that may be useful for other
clients of the HTML renderer.

R=jasvir@gmail.com