make html-sanitizer more standalone friendly http://codereview.appspot.com/4432045
this addresses http://code.google.com/p/google-caja/issues/detail?id=1309

1. html-sanitizer-minified.js includes css-defs.js but doesn't
use it.  removing css-defs reduces html-sanitizer-minified about 60%.

2. some examples include both domita-minified.js and
html-sanitizer-minified.js, which is redundant since domita-minified.js
includes html-sanitizer.  this change removes the redundancy.

3. html-sanitizer is not a self-contained module.  in particular,
it's difficult to rename the 'html' object it defines, because it
includes some functions that refer to itself using the 'html'
global name, which is unnecessary.  this changes the functions
to use local references instead.

4. html-sanitizer also uses the global 'html4'.  this changes
it to capture the value of html4 at definition time.

5. capturing 'html4' means that html4-defs.js needs to be included
before html-sanitizer.js.  there are a few instances that didn't,
and this fixes that.

R=jasvir@gmail.com